Made the screen not shut off
I made the objects spawn along any possible y coordinate instead of 400
plus any possible y coordinate

Instead of having the game resume with a brand new game in onResume(), I
made a boolean variable that keeps track of whether or not we are
initializing the activity or we have paused it. If we paused it, just
start the game loop up again. Otherwise, create a new one and then start
it.